Reading remote FTP directory entries

+++ test/source/network/test-set/ftp_test_set.e
@@ -6,8 +6,8 @@ note
 	contact: "finnian at eiffel hyphen loop dot com"
 
 	license: "MIT license (See: en.wikipedia.org/wiki/MIT_License)"
-	date: "2023-08-19 14:49:43 GMT (Saturday 19th August 2023)"
-	revision: "9"
+	date: "2023-08-22 12:20:13 GMT (Tuesday 22nd August 2023)"
+	revision: "10"
 
 class
 	FTP_TEST_SET
@@ -17,6 +17,15 @@ inherit
 		undefine
 			new_lio
 		end
+
+	EL_PLAIN_TEXT_FILE_STATE_MACHINE
+		undefine
+			default_create
+		redefine
+			make
+		end
+
+	EL_MODULE_ARGS
 
 	EL_CRC_32_TESTABLE
 
@@ -31,51 +40,50 @@ feature {NONE} -- Initialization
 		-- initialize `test_table'
 		do
 			make_named (<<
-				["ftp_directory_exists", agent test_ftp_directory_exists],
-				["reuse_authenticator",	 agent test_reuse_authenticator]
+				["ftp_directory_exists", agent test_ftp_directory_exists]
 			>>)
 		end
 
 feature -- Tests
 
 	test_ftp_directory_exists
-		-- FILE_TEST_SET.test_ftp_directory_exists
+		-- FTP_TEST_SET.test_ftp_directory_exists
 		local
-			config: EL_FTP_CONFIGURATION
-			ftp: EL_FTP_PROTOCOL; home_dir: DIR_PATH
-			url: FTP_URL; dir_path: EL_DIR_PATH
+			config: EL_FTP_CONFIGURATION ftp: EL_FTP_PROTOCOL; dir_path: EL_DIR_PATH
 		do
---			create config.
+			config := new_pyxis_config.ftp
+			if attached Args.value ("pp") as pp and then pp.count > 0 then
+				config.authenticate (pp)
+			else
+				config.authenticate (Void)
+			end
 
---			create ftp.make_write ([url.host, home_dir])
-			ftp.open
-			ftp.login
-			assert ("logged in", ftp.is_logged_in)
-			ftp.change_home_dir
-			dir_path := "test_location"
-			ftp.make_directory (dir_path)
-			assert ("directory exists", ftp.directory_exists (dir_path))
-			ftp.remove_directory (dir_path)
-			assert ("not directory exists", not ftp.directory_exists (dir_path))
-			ftp.close
-		end
-
-	test_reuse_authenticator
-		local
-			config: EL_FTP_CONFIGURATION; item: EL_FTP_UPLOAD_ITEM
-			ftp: EL_FTP_PROTOCOL
-		do
-			create config.make_default
-			config.authenticate
-			across file_list as list loop
+			if config.credential.is_valid then
 				create ftp.make_write (config)
+				ftp.open
 				ftp.login
-				ftp.make_directory (Test_set)
-				lio.put_path_field ("Uploading", list.item)
-				lio.put_new_line
-				create item.make (list.item, Test_set)
-				ftp.upload (item)
+				assert ("logged in", ftp.is_logged_in)
+				ftp.change_home_dir
+				dir_path := "W_code/C1"
+				ftp.make_directory (dir_path)
+				assert ("directory exists", ftp.directory_exists (dir_path))
+
+				from until dir_path.is_empty loop
+					ftp.remove_directory (dir_path)
+					assert ("not directory exists", not ftp.directory_exists (dir_path))
+					dir_path := dir_path.parent
+				end
+
+				across ftp.file_list ("css") as list loop
+					lio.put_path_field ("css %S", list.item)
+					lio.put_new_line
+				end
+				ftp.read_file_count ("css")
+				assert ("4 entries", ftp.file_count = 4)
+
 				ftp.close
+			else
+				failed ("Authenticated")
 			end
 		end
 
@@ -84,6 +92,11 @@ feature {NONE} -- Implementation
 	source_file_list: EL_FILE_PATH_LIST
 		do
 			Result := OS.file_list (Data_dir, "*.txt")
+		end
+
+	new_pyxis_config: EL_PYXIS_FTP_CONFIGURATION
+		do
+			create Result.make (Dev_environ.Eiffel_loop_dir + "doc-config/config.pyx")
 		end
 
 feature {NONE} -- Constants
